build: Move XDNA_BIN_DIR into CMAKE_BINARY_DIR.
When setting a CMAKE_INSTALL_PREFIX, the build tried to install binaries
into `/bins`, the system root directory, which luckily fails due to
permissions.

It is a very bad practice to install into absolute paths and creating
directories in the system root using the build system.

In addition `XDNA_BIN_DIR` was already prefixed with `CMAKE_BINARY_DIR`
in several places of the build system. This patch makes the manual
prefixing redundant and adds consistency.

Signed-off-by: Lubosz Sarnecki <lubosz@gmail.com>